Admin dashboard
A dashboard for Ubuntu hub website that allows admins to view the list of newly added products in the form of a table that is dynamically updated through the use of JavaScript as soon as new products are added. The board contains cards that serve as an overview that enables the admin to check the number of massages as well as products in the inventory.